// frontend/new-web/src/components/Crud/types.ts
export type IObject = Record<string, any>;

/**
 * 组件类型定义
 */
type DateComponent = "date-picker" | "time-picker" | "time-select" | "custom-tag" | "input-tag";
type InputComponent = "input" | "select" | "input-number" | "cascader" | "tree-select";
type OtherComponent =
  | "text"
  | "radio"
  | "checkbox"
  | "switch"
  | "rate"
  | "slider"
  | "icon-select"
  | "custom";
export type ISearchComponent =
  | DateComponent
  | InputComponent
  | "radio"
  | "checkbox"
  | "switch"
  | "rate"
  | "slider"
  /** 在 customComponents 中注册的自定义搜索控件类型 */
  | "user-table-select";
export type IComponentType = DateComponent | InputComponent | OtherComponent;